Use store block data; introduce common package
