+++ pkg/rpc/core/utils_test.go
@@ -4,6 +4,7 @@ import (
 	"crypto/sha256"
 	"encoding/hex"
 	"fmt"
+	"github.com/stretchr/testify/require"
 	"strings"
 	"testing"
 
@@ -21,7 +22,7 @@ func TestTruncateNodeID(t *testing.T) {
 		expectedPrefix := fullNodeIDStr[:NodeIDByteLength*2] // First 20 bytes -> 40 chars
 
 		truncatedNodeID, err := TruncateNodeID(fullNodeIDStr)
-		assert.NoError(t, err, "TruncateNodeID should not return an error for a valid ID")
+		require.NoError(t, err, "TruncateNodeID should not return an error for a valid ID")
 		assert.Equal(t, NodeIDByteLength*2, len(truncatedNodeID), fmt.Sprintf("Truncated Node ID should be %d characters long", NodeIDByteLength*2))
 		assert.Equal(t, expectedPrefix, truncatedNodeID, "Truncated Node ID should be the first 20 bytes of the original ID")
 	})
@@ -37,12 +38,9 @@ func TestTruncateNodeID(t *testing.T) {
 	// Test case 3: Invalid hex string
 	t.Run("invalid hex string", func(t *testing.T) {
 		invalidHexStr := "not-a-hex-string-at-all-and-should-be-long-enough-to-pass-length-check"
-		// Ensure it would be long enough if it were hex
-		if len(invalidHexStr) < NodeIDByteLength*2 {
-			invalidHexStr = strings.Repeat("x", NodeIDByteLength*2) // Make it long enough but still invalid hex
-		}
+		require.GreaterOrEqual(t, len(invalidHexStr), NodeIDByteLength*2)
 		_, err := TruncateNodeID(invalidHexStr)
-		assert.Error(t, err, "TruncateNodeID should return an error for an invalid hex string")
+		require.Error(t, err, "TruncateNodeID should return an error for an invalid hex string")
 		assert.Contains(t, err.Error(), "failed to decode node ID", "Error message should indicate hex decoding failure")
 	})
 
@@ -56,7 +54,7 @@ func TestTruncateNodeID(t *testing.T) {
 	t.Run("exact length node ID", func(t *testing.T) {
 		exactLengthNodeIDStr := strings.Repeat("a", NodeIDByteLength*2) // 40 chars
 		truncatedNodeID, err := TruncateNodeID(exactLengthNodeIDStr)
-		assert.NoError(t, err, "TruncateNodeID should not return an error for an ID of exact length")
+		require.NoError(t, err, "TruncateNodeID should not return an error for an ID of exact length")
 		assert.Equal(t, exactLengthNodeIDStr, truncatedNodeID, "Truncated Node ID should be the same as input for exact length")
 	})
 }
